requirements.txt
pandas
confluent-kafka
scikit-learn
joblib
pyspark
pyarrow
fastparquet
kafka-python
